Refactor: Migrate window frame to common module
This commit migrates the window frame implementation to the common module, enabling platform-specific customizations.

The `TabViewWindowContent` composable is introduced to handle window content with padding insets, header, and footer.

The `WindowFrame` composable is added to the common module and delegates to platform-specific implementations: `WindowsWindowFrame` for Windows and `MacOSWindowFrame` for macOS.

The `TabViewScreen` is updated to use the new `WindowFrame` composable, passing the appropriate parameters for each platform.
